Fixing the type detection of Bluetooth Uuids, and fixing the 32bit Uuid to UInt32 conversion.
If i interpret the Bluetooth spec correctly, a 128bit Uuid is any Uuid that is not based on the base Uuid.
To detect if the type of a Uuid is 16 or 32bit, the Uuid must be based on the base Uuid, and then we check the MSBbytes for zeros.
The conversion of a 32bit Uuid to UInt32 mixed up the byte order.
Unit tests are added as proof for the fixes.
runsettings are updated as the unit tests require real hardware.

Signed-off-by: Serge Gommers <serge.gommers@hotmail.com>